Precision too

+++ apps/webserver/src/index.js
@@ -29,6 +29,10 @@ app.use(bodyParser.json({
     limit : '100kb'
 }));
 
+app.use(bodyParser.urlencoded({
+    extended: true
+}));
+
 // busboy for multipart
 app.use(busboy());
 
@@ -46,27 +50,31 @@ app.get('/', (req, res) => {
 app.post('/upload', (req, res, next) => {
     let fstream;
     req.pipe(req.busboy);
-    req.busboy.on('file', (field, file, fileName) => {
-        console.info('Uploading', fileName);
-        fstream = fs.createWriteStream(`${__dirname}/static/input/${fileName}`);
-        file.pipe(fstream);
-        fstream.on('close', () => {
-            console.info('Uploaded', fileName);
-            child.exec(`primitive -i ./dist/static/input/${fileName} -o ./dist/static/output/${fileName} -n 100`, (err, stdout, stderr) => {
-                if(!err) {
-                    fs.emptyDir(`${__dirname}/static/input`, (err) => {
+
+    req.busboy.on('field', (fieldName, value) => {
+        if(fieldName === 'precision') {
+            req.busboy.on('file', (field, file, fileName) => {
+                fstream = fs.createWriteStream(`${__dirname}/static/input/${fileName}`);
+                file.pipe(fstream);
+                fstream.on('close', () => {
+                    console.info('Uploaded', fileName);
+                    child.exec(`primitive -i ./dist/static/input/${fileName} -o ./dist/static/output/${fileName} -n ${value}`, (err, stdout, stderr) => {
                         if(!err) {
-                            res.redirect(`/output/${fileName}`);
-                            console.log(stdout);
-                        }else {
+                            fs.emptyDir(`${__dirname}/static/input`, (err) => {
+                                if(!err) {
+                                    res.redirect(`/output/${fileName}`);
+                                    console.log(stdout);
+                                }else {
+                                    console.error(err);
+                                }
+                            })
+                        } else {
                             console.error(err);
                         }
-                    })
-                } else {
-                    console.error(err);
-                }
+                    });
+                });
             });
-        });
+        }
     });
 });
